pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kube-ovn-checker"
version = "1.0.0"
description = "基于 LLM 的 Kube-OVN 网络问题智能诊断工具"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Kube-OVN Community"}
]
keywords = ["kubernetes", "kube-ovn", "networking", "diagnostics", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Networking",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "langgraph>=1.0.3",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "filelock>=3.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "msgpack>=1.0.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
kube-ovn-checker = "kube_ovn_checker.cli.main:main"

[project.urls]
Homepage = "https://github.com/kube-ovn/kube-ovn-langgraph-checker"
Repository = "https://github.com/kube-ovn/kube-ovn-langgraph-checker"
Documentation = "https://kubeovn.github.io/docs/"
"Bug Tracker" = "https://github.com/kube-ovn/kube-ovn-langgraph-checker/issues"

[tool.setuptools]
package-data = {kube_ovn_checker = ["knowledge/*.md", "knowledge/crds/*.md", "knowledge/issues/*.md"]}

[tool.setuptools.packages.find]
exclude = ["tests*", "docs*", "build*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
